rename, refactor & improve out of scope question generator
