+++ giskard/rag/question_generators/__init__.py
@@ -3,7 +3,7 @@ from .complex_questions import ComplexQuestionsGenerator, complex_questions
 from .conversational_questions import ConversationalQuestionsGenerator, conversational_questions
 from .distracting_questions import DistractingQuestionsGenerator, distracting_questions
 from .double_questions import DoubleQuestionsGenerator, double_questions
-from .ookb_questions import OutOfKnowledgeBaseGenerator, ookb_questions
+from .oos_questions import OutOfScopeGenerator, oos_questions
 from .question_types import COMPONENT_DESCRIPTIONS, QUESTION_ATTRIBUTION, RAGComponents
 from .simple_questions import SimpleQuestionsGenerator, simple_questions
 from .situational_questions import SituationalQuestionsGenerator, situational_questions
@@ -19,12 +19,12 @@ __all__ = [
     "DistractingQuestionsGenerator",
     "SituationalQuestionsGenerator",
     "DoubleQuestionsGenerator",
-    "OutOfKnowledgeBaseGenerator",
+    "OutOfScopeGenerator",
     "simple_questions",
     "complex_questions",
     "conversational_questions",
     "distracting_questions",
     "situational_questions",
     "double_questions",
-    "ookb_questions",
+    "oos_questions",
 ]
